% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colbarmat}
\alias{colbarmat}
\title{Example color annotation for columns heatmap demonstration}
\format{A character matrix with 2 color rows and 20 columns:}
\usage{
colbarmat
}
\description{
Example color annotation for columns heatmap demonstration
}
\keyword{datasets}

