% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colbars1}
\alias{colbars1}
\title{A colorbar example for columns}
\format{A vector for colors of 20 columns:}
\usage{
colbars1
}
\description{
A colorbar example for columns
}
\keyword{datasets}

