% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat}
\alias{mat}
\title{Example data for heatmap demonstration}
\format{A matrix with 50 rows and 20 columns:}
\usage{
mat
}
\description{
A randomly generated numeric matrix for heatmap con continous scale
}
\keyword{datasets}
