% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rowbarmat}
\alias{rowbarmat}
\title{Example color annotation for rows heatmap demonstration}
\format{A character matrix with 50 rows and 2 color columns:}
\usage{
rowbarmat
}
\description{
Example color annotation for rows heatmap demonstration
}
\keyword{datasets}
